Final Checkpoint 2 "Completed Results" Build
This is the final build of the second checkpoint of the Practice Databases Assignment, which has taken place after  the Video titled "Star Ratings Part 2" in the ePub document associated with this achievement standard. Next, I will begin on Checkpoint 3, which is in regard to the sarching of the database.