Practice Databases Project after "Basic Search Part 3" Video
This is the build of my Practice Databases Project after the 'Basic Search Part 3' Video, and ultimately the 'Basic Search' section as a whole. This version of the program has implemented an 'App Name and/or Developer Name' search feature, which is now functional. The nect step of this Checkpoint is to implement a 'button' that shows results that are 'Free with no In-App Purchases', which will therefore conclude the work associated with this checkpoint.