Practice Databases Project after "Checkpoint 3 - Basic Search"
This is the build of the Practice Databases code after the third Chekpoint of this poject titles "Basic Search". As of this poin in the project's development, both search functions are functional (however not fully polished), and screenshots of this version have been turned in as exidence of this Project's Checkpont 3 being completed.