Practice Databases Project after "Advanced Search Part Testing" Video
This is the build of the Practice Databases project that has now both implemented and error corrected various sections of the "Advanced Search" form as part of this database. This code may not be completely functional and error corrected, however it has been completed up until the 'ePub' video mentioned above.