Practice Databases Project after "Entry Form Setup Part 3" Video
This is the build of the Practice Databases Project that, while some errors do still exist within it, is completed until the end of the afforementioned video, which also signifies the end of the "Entry Form Setup" Section of the assigned 'ePub' Document. The next step forward for this code will be the insertion of new items into the Project's Database from the website's afforementioned 'Form'.