Practice Databases Project after "Get Values from Form" Video
This is the build of this project that improves the database website's 'Add an App' page, which now holds data once it has been submittend, most likely among other improvements. The next stage of development for this project may be to implement the content under the heading 'Add Entries' in the relevant, provided, ePub document.